requirements.txt
aiogram==2.9.2
python-dotenv